+++ diesel/src/expression/mod.rs
@@ -35,6 +35,8 @@ pub mod functions;
 pub mod grouped;
 pub mod helper_types;
 #[doc(hidden)]
+pub mod nullable;
+#[doc(hidden)]
 pub mod ordering;
 #[doc(hidden)]
 pub mod predicates;